fix major bug when optimization is on
When the payload is smaller than cell size (that is when
sizeof(PAYLOAD)<2*sizeof(INDEX)) the number of allocated cells was wrong
on odd number of requested element, causing writing beyond allocated
memory and corruption of management linked lists.
There is a new strategy to manage deallocated regions: there are put in
order in the linked list and possibly merged with adjacent regions.